service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.48.3",
    "fakeredis>=2.34.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "mongomock>=4.3.0",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "taskiq>=0.12.1",
    "taskiq-fastapi>=0.4.0",
    "taskiq-redis>=1.2.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "requests>=2.32.5",
    "ruff>=0.14.7",
    "taskiq[reload]>=0.12.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"
preview = true

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "C90",   # mccabe
    "I",     # isort
    "ASYNC", # flake8-async
    "B",     # flake8-bugbear
    "BLE",   # flake8-blind-except
    "COM",   # flake8-commas
    "C4",    # flake8-comprehensions
    "DTZ",   # flake8-datetimez
    "EM",    # flake8-errmsg
    "ICN",   # flake8-import-conventions
    "PYI",   # flake8-pyi
    "Q",     # flake8-quotes
    "SIM",   # flake8-simplify
    "TCH",   # flake8-typechecking
    "PTH",   # flake8-use-pathlib
    "UP",    # pyupgrade
    "RUF",   # ruff-specific
    "TRY",   # tryceratops
    "FURB",  # refurb
]
ignore = ["COM812"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["--cov-report=term", "--cov=src"]
pythonpath = ["src", "."]
markers = ["unit: unit tests", "integration: integration tests"]

[tool.coverage.run]
omit = ["*/dependency_injector/*"]
